requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
class-registry==2.1.2
Django==4.2.4
django-class-auth-views==1.0.2
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-api-tracking==1.8.4
filters==1.3.2
filters-django==1.0.5
psycopg==3.1.10
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
regex==2023.8.8
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
